moving ariaLive into ui/ dir
https://coveord.atlassian.net/browse/JSUI-2237
