+++ src/ui/SearchInterface/SearchInterface.ts
@@ -44,7 +44,7 @@ import { MEDIUM_SCREEN_WIDTH, ResponsiveComponents, SMALL_SCREEN_WIDTH } from '.
 import { FacetColumnAutoLayoutAdjustment } from './FacetColumnAutoLayoutAdjustment';
 import { FacetValueStateHandler } from './FacetValueStateHandler';
 import RelevanceInspectorModule = require('../RelevanceInspector/RelevanceInspector');
-import { AriaLive } from '../../utils/AriaLive';
+import { AriaLive } from '../AriaLive/AriaLive';
 
 export interface ISearchInterfaceOptions {
   enableHistory?: boolean;
