+++ unitTests/Test.ts
@@ -645,7 +645,7 @@ MagicBoxClearTest();
 import { AccessibleButtonTest } from './utils/AccessibleButtonTest';
 AccessibleButtonTest();
 
-import { AriaLiveTest } from './utils/AriaLiveTest';
+import { AriaLiveTest } from './ui/AriaLiveTest';
 AriaLiveTest();
 
 import { QuerySummaryUtilsTest } from './utils/QuerySummaryUtilsTest';
